unreal-codex-agent/app/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
websockets==12.0
requests==2.31.0
openai>=1.30.0
groq>=0.8.0
google-generativeai>=0.7.0
anthropic>=0.28.0
